pyproject.toml
[tool.poetry]
name = "athenaeum_back"
version = "0.1.0"
description = "Book Sharing App"
authors = ["Code Wizard <codewizard001@gmail.com>"]

[tool.poetry.dependencies]
python = ">3.7"
django = "^3.0.7"
passlib = "^1.7.2"
mongoengine = "^0.20.0"
djangorestframework = "^3.11.0"
django-rest-framework-mongoengine = "^3.4.1"
django-cors-headers = "^3.3.0"
django-cors-middleware = "^1.5.0"
schema = "^0.7.2"
pycountry = "^19.8.18"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
